done states finally

+++ source/states/TitleState.hx
@@ -37,32 +37,39 @@ typedef TitleData =
 
 class TitleState extends MusicBeatState
 {
+{
 	public static var muteKeys:Array<FlxKey> = [FlxKey.ZERO];
 	public static var volumeDownKeys:Array<FlxKey> = [FlxKey.NUMPADMINUS, FlxKey.MINUS];
 	public static var volumeUpKeys:Array<FlxKey> = [FlxKey.NUMPADPLUS, FlxKey.PLUS];
 
 	public static var initialized:Bool = false;
 
-	var credGroup:FlxGroup = new FlxGroup();
-	var textGroup:FlxGroup = new FlxGroup();
-	var blackScreen:FlxSprite;
-	var credTextShit:Alphabet;
-	var ngSpr:FlxSprite;
+	public var credGroup:FlxGroup = new FlxGroup();
+	public var textGroup:FlxGroup = new FlxGroup();
+	public var blackScreen:FlxSprite;
+	public var credTextShit:Alphabet;
+	public var ngSpr:FlxSprite;
 	
-	var titleTextColors:Array<FlxColor> = [0xFF33FFFF, 0xFF3333CC];
-	var titleTextAlphas:Array<Float> = [1, .64];
+	public var titleTextColors:Array<FlxColor> = [0xFF33FFFF, 0xFF3333CC];
+	public var titleTextAlphas:Array<Float> = [1, .64];
 
-	var curWacky:Array<String> = [];
+	public var curWacky:Array<String> = [];
 
-	var wackyImage:FlxSprite;
+	public var wackyImage:FlxSprite;
 
 	#if TITLE_SCREEN_EASTER_EGG
-	final easterEggKeys:Array<String> = [
+	public final easterEggKeys:Array<String> = [
 		'SHADOW', 'RIVEREN', 'BBPANZU', 'PESSY'
 	];
-	final allowedKeys:String = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
-	var easterEggKeysBuffer:String = '';
+	public final allowedKeys:String = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
+	public var easterEggKeysBuffer:String = '';
 	#end
+
+	public var logoBl:FlxSprite;
+	public var gfDance:FlxSprite;
+	public var danceLeft:Bool = false;
+	public var titleText:FlxSprite;
+	public var swagShader:ColorSwap = null;
 
 	override public function create():Void
 	{
@@ -75,6 +82,8 @@ class TitleState extends MusicBeatState
 			ClientPrefs.loadPrefs();
 			Language.reloadPhrases();
 		}
+
+		setStateScript();
 
 		curWacky = FlxG.random.getObject(getIntroTextShit());
 
@@ -124,6 +133,8 @@ class TitleState extends MusicBeatState
 		persistentUpdate = true;
 		if (!initialized && FlxG.sound.music == null)
 			FlxG.sound.playMusic(Paths.music('freakyMenu'), 0);
+
+		StateScriptHandler.callOnScripts("onStartIntro", []);
 
 		loadJsonData();
 		#if TITLE_SCREEN_EASTER_EGG easterEggData(); #end
@@ -210,23 +221,23 @@ class TitleState extends MusicBeatState
 		else
 			initialized = true;
 
-		// credGroup.add(credTextShit);
+		StateScriptHandler.callOnScripts("onStartIntroPost", []);
 	}
 
 	// JSON data
-	var characterImage:String = 'gfDanceTitle';
-	var animationName:String = 'gfDance';
+	public var characterImage:String = 'gfDanceTitle';
+	public var animationName:String = 'gfDance';
 
-	var gfPosition:FlxPoint = FlxPoint.get(512, 40);
-	var logoPosition:FlxPoint = FlxPoint.get(-150, -100);
-	var enterPosition:FlxPoint = FlxPoint.get(100, 576);
+	public var gfPosition:FlxPoint = FlxPoint.get(512, 40);
+	public var logoPosition:FlxPoint = FlxPoint.get(-150, -100);
+	public var enterPosition:FlxPoint = FlxPoint.get(100, 576);
 	
-	var useIdle:Bool = false;
-	var musicBPM:Float = 102;
-	var danceLeftFrames:Array<Int> = [15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29];
-	var danceRightFrames:Array<Int> = [30, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14];
+	public var useIdle:Bool = false;
+	public var musicBPM:Float = 102;
+	public var danceLeftFrames:Array<Int> = [15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29];
+	public var danceRightFrames:Array<Int> = [30, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14];
 
-	function loadJsonData()
+	public function loadJsonData()
 	{
 		if(Paths.fileExists('images/gfDanceTitle.json', TEXT))
 		{
@@ -263,7 +274,7 @@ class TitleState extends MusicBeatState
 		//else trace('[WARN] No Title JSON detected, using default values.');
 	}
 
-	function easterEggData()
+	public function easterEggData()
 	{
 		if (FlxG.save.data.psychDevsEasterEgg == null) FlxG.save.data.psychDevsEasterEgg = ''; //Crash prevention
 		var easterEgg:String = FlxG.save.data.psychDevsEasterEgg;
@@ -298,7 +309,7 @@ class TitleState extends MusicBeatState
 		}
 	}
 
-	function getIntroTextShit():Array<Array<String>>
+	public function getIntroTextShit():Array<Array<String>>
 	{
 		#if MODS_ALLOWED
 		var firstArray:Array<String> = Mods.mergeAllTextsNamed('data/introText.txt');
@@ -316,11 +327,11 @@ class TitleState extends MusicBeatState
 		return swagGoodArray;
 	}
 
-	var transitioning:Bool = false;
+	public var transitioning:Bool = false;
 	private static var playJingle:Bool = false;
 	
-	var newTitle:Bool = false;
-	var titleTimer:Float = 0;
+	public var newTitle:Bool = false;
+	public var titleTimer:Float = 0;
 
 	override function update(elapsed:Float)
 	{
@@ -452,7 +463,7 @@ class TitleState extends MusicBeatState
 		super.update(elapsed);
 	}
 
-	function createCoolText(textArray:Array<String>, ?offset:Float = 0)
+	public function createCoolText(textArray:Array<String>, ?offset:Float = 0)
 	{
 		for (i in 0...textArray.length)
 		{
@@ -467,7 +478,7 @@ class TitleState extends MusicBeatState
 		}
 	}
 
-	function addMoreText(text:String, ?offset:Float = 0)
+	public function addMoreText(text:String, ?offset:Float = 0)
 	{
 		if(textGroup != null && credGroup != null) {
 			var coolText:Alphabet = new Alphabet(0, 0, text, true);
@@ -478,7 +489,7 @@ class TitleState extends MusicBeatState
 		}
 	}
 
-	function deleteCoolText()
+	public function deleteCoolText()
 	{
 		while (textGroup.members.length > 0)
 		{
@@ -489,6 +500,7 @@ class TitleState extends MusicBeatState
 
 	private var sickBeats:Int = 0; //Basically curBeat but won't be skipped if you hold the tab or resize the screen
 	public static var closedState:Bool = false;
+	public var allowCustomIntro:Bool = false;
 	override function beatHit()
 	{
 		super.beatHit();
@@ -512,49 +524,50 @@ class TitleState extends MusicBeatState
 		if(!closedState)
 		{
 			sickBeats++;
-			switch (sickBeats)
-			{
-				case 1:
-					//FlxG.sound.music.stop();
-					FlxG.sound.playMusic(Paths.music('freakyMenu'), 0);
-					FlxG.sound.music.fadeIn(4, 0, 0.7);
-				case 2:
-					createCoolText(['Psych Engine by'], 40);
-				case 4:
-					addMoreText('Shadow Mario', 40);
-					addMoreText('Riveren', 40);
-				case 5:
-					deleteCoolText();
-				case 6:
-					createCoolText(['Not associated', 'with'], -40);
-				case 8:
-					addMoreText('newgrounds', -40);
-					ngSpr.visible = true;
-				case 9:
-					deleteCoolText();
-					ngSpr.visible = false;
-				case 10:
-					createCoolText([curWacky[0]]);
-				case 12:
-					addMoreText(curWacky[1]);
-				case 13:
-					deleteCoolText();
-				case 14:
-					addMoreText('Friday');
-				case 15:
-					addMoreText('Night');
-				case 16:
-					addMoreText('Funkin'); // credTextShit.text += '\nFunkin';
+			if (!allowCustomIntro) {
+				switch (sickBeats)
+				{
+					case 1:
+						//FlxG.sound.music.stop();
+						FlxG.sound.playMusic(Paths.music('freakyMenu'), 0);
+						FlxG.sound.music.fadeIn(4, 0, 0.7);
+					case 2:
+						createCoolText(['Forreal Engine by'], 40);
+					case 4:
+						addMoreText('Huy1234TH', 40);
+					case 5:
+						deleteCoolText();
+					case 6:
+						createCoolText(['Not associated', 'with'], -40);
+					case 8:
+						addMoreText('newgrounds', -40);
+						ngSpr.visible = true;
+					case 9:
+						deleteCoolText();
+						ngSpr.visible = false;
+					case 10:
+						createCoolText([curWacky[0]]);
+					case 12:
+						addMoreText(curWacky[1]);
+					case 13:
+						deleteCoolText();
+					case 14:
+						addMoreText('Friday');
+					case 15:
+						addMoreText('Night');
+					case 16:
+						addMoreText('Funkin'); // credTextShit.text += '\nFunkin';
 
-				case 17:
-					skipIntro();
-			}
+					case 17:
+						skipIntro();
+				}
+			} else {} // own custom intro eyy
 		}
 	}
 
-	var skippedIntro:Bool = false;
-	var increaseVolume:Bool = false;
-	function skipIntro():Void
+	public var skippedIntro:Bool = false;
+	public var increaseVolume:Bool = false;
+	public function skipIntro():Void
 	{
 		if (!skippedIntro)
 		{
